requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
pydantic==2.10.5
requests==2.32.3
python-dotenv==1.0.1
